refactor: extract useIsMobile hook to shared module
